HardwareMap
HardwareMap is almost completed, just need to connect to the robot to fix the .get command so we can sync android studio with the robot
